+++ sdk/core/azure-core/azure/core/credentials.py
@@ -4,9 +4,11 @@
 # license information.
 # -------------------------------------------------------------------------
 from collections import namedtuple
-from typing import Any, NamedTuple, Optional
+from typing import Any, NamedTuple, Optional, TypeVar, Tuple, Union
 from typing_extensions import Protocol, runtime_checkable
 
+
+TKeys = TypeVar('TKeys', bound=Union[str, Tuple[str, ...]])
 
 class AccessToken(NamedTuple):
     """Represents an OAuth access token."""
@@ -59,36 +61,36 @@ class AzureKeyCredential:
     """Credential type used for authenticating to an Azure service.
     It provides the ability to update the key without creating a new client.
 
-    :param str key: The key used to authenticate to an Azure service
+    :param Union[str, Tuple[str, ...]] key: The key used to authenticate to an Azure service
     :raises: TypeError
     """
 
-    def __init__(self, key: str) -> None:
-        if not isinstance(key, str):
-            raise TypeError("key must be a string.")
+    def __init__(self, key: TKeys) -> None:
+        if not isinstance(key, str) and not isinstance(key, Tuple):
+            raise TypeError("key must be a string or a tuple of strings.")
         self._key = key
 
     @property
-    def key(self) -> str:
+    def key(self) -> Union[str, Tuple[str, ...]]:
         """The value of the configured key.
 
-        :rtype: str
+        :rtype: Union[str, Tuple[str, ...]]
         """
         return self._key
 
-    def update(self, key: str) -> None:
+    def update(self, key: Union[str, Tuple[str, ...]]) -> None:
         """Update the key.
 
         This can be used when you've regenerated your service key and want
         to update long-lived clients.
 
-        :param str key: The key used to authenticate to an Azure service
+        :param Union[str, Tuple[str, ...]] key: The key used to authenticate to an Azure service
         :raises: ValueError or TypeError
         """
         if not key:
             raise ValueError("The key used for updating can not be None or empty")
-        if not isinstance(key, str):
-            raise TypeError("The key used for updating must be a string.")
+        if not isinstance(key, str) and not isinstance(key, Tuple):
+            raise TypeError("The key used for updating must be a string or a tuple of strings.")
         self._key = key